Add in div grid and size functionality
This commit adds a lot of functionality for not much code! First,
a grid was created using loops and various DOM changing elements.

Then, the ability to alter the size of the grid was added. But the
crucial part was ensuring that the size of the box remains the
same, whilst the number of pixels increases. This was achieved
using a ${} placeholder with the number of 'px' coming at the end
(outside of the placeholder brackets).